refactor: CheckCoinReturn.js 에서 coinReturn 관련 로직 분리
- SetCoinReturn 에서 반환하기 버튼 클릭 이후 로직 관리
