+++ src/controller/CheckCoinReturn.js
@@ -1,4 +1,5 @@
-import { DOM, LOCAL_STORAGE, EVENT, TEMPLATE, ERROR_MESSAGE, NUMBER, STRING, COIN } from '../utils/constant.js';
+import { DOM, EVENT, ERROR_MESSAGE, NUMBER } from '../utils/constant.js';
+import SetCoinReturn from './SetCoinReturn.js';
 import SetPurchaseButtons from './SetPurchaseButtons.js';
 
 export default class CheckCoinReturn {
@@ -24,96 +25,20 @@ export default class CheckCoinReturn {
     $$purchaseButtons.forEach(($purchaseButton) => this.onClickPurchaseButton($purchaseButton));
   };
 
-  spendCoin = (vendingMachineCoinQuantity, currentCoin, vendingMachineCoinAmount, userCoinAmount) => {
-    while (vendingMachineCoinQuantity > NUMBER.ZERO && userCoinAmount - currentCoin >= NUMBER.ZERO) {
-      userCoinAmount -= currentCoin;
-      vendingMachineCoinAmount -= currentCoin;
-      vendingMachineCoinQuantity -= NUMBER.ONE;
-    }
-
-    return [vendingMachineCoinQuantity, vendingMachineCoinAmount, userCoinAmount];
-  };
-
-  compareVendingMachineCoinAndUserCoin = (vendingMachineCoinQuantityHash, vendingMachineCoinAmount, userCoinAmount) => {
-    COIN.LIST.forEach((currentCoin) => {
-      const [spendVendingMachineCoinQuantity, spendVendingMachineCoinAmount, spendUserCoinAmount] = this.spendCoin(
-        vendingMachineCoinQuantityHash[currentCoin],
-        currentCoin,
-        vendingMachineCoinAmount,
-        userCoinAmount
-      );
-      vendingMachineCoinQuantityHash[currentCoin] -= spendVendingMachineCoinQuantity;
-      vendingMachineCoinAmount = spendVendingMachineCoinAmount;
-      userCoinAmount = spendUserCoinAmount;
-    });
-
-    return [vendingMachineCoinQuantityHash, vendingMachineCoinAmount, userCoinAmount];
-  };
-
-  renderReturnCoin = (vendingMachineCoinQuantityHash) => {
-    this.render.returnCoinTemplate(vendingMachineCoinQuantityHash);
-  };
-
-  hasReturnCoin = () => {
-    const [vendingMachineCoinQuantityHash, vendingMachineCoinAmount, userCoinAmount] =
-      this.compareVendingMachineCoinAndUserCoin(
-        { ...this.coins.getCoinsHash() },
-        this.coins.getCoinAmount(),
-        this.vendingMachine.getChargeAmount()
-      );
-
-    this.coins.replaceCoinsHash(vendingMachineCoinQuantityHash);
-    this.coins.replaceCoinAmount(vendingMachineCoinAmount);
-    this.vendingMachine.replaceChargeAmount(userCoinAmount);
-    this.renderReturnCoin(vendingMachineCoinQuantityHash);
-  };
-
-  renderPurchaseChargeAmount = () => {
-    this.render.purchaseChargeAmountTemplate();
-  };
-
-  replaceManageChargeAmount = (previousCoinAmount, currentCoinAmount) => {
-    this.vendingMachineManageMenuTemplate = this.vendingMachineManageMenuTemplate.replace(
-      TEMPLATE.VENDING_MACHINE_CHARGE_AMOUNT(previousCoinAmount + STRING.WON),
-      TEMPLATE.VENDING_MACHINE_CHARGE_AMOUNT(currentCoinAmount + STRING.WON)
-    );
-  };
-
-  replaceManageCoin = (previousCoinEntries, currentCoinEnries) => {
-    previousCoinEntries.forEach((previousCoinEntry, index) => {
-      const [previousCoin, previousCoinQuantity] = previousCoinEntry;
-      const [currentCoin, currentCoinQuantity] = currentCoinEnries[index];
-
-      this.vendingMachineManageMenuTemplate = this.vendingMachineManageMenuTemplate.replace(
-        TEMPLATE.VENDING_MACHINE_COIN(previousCoin, previousCoinQuantity),
-        TEMPLATE.VENDING_MACHINE_COIN(currentCoin, currentCoinQuantity)
-      );
-    });
-  };
-
-  reRenderMachineManageMenu = (previousCoinEntries, previousCoinAmount) => {
-    this.vendingMachineManageMenuTemplate = localStorage.getItem(LOCAL_STORAGE.VENDING_MACHINE_MANAGE_MENU);
-    const currentCoinEnries = Object.entries(this.coins.getCoinsHash());
-    const currentCoinAmount = this.coins.getCoinAmount();
-
-    this.replaceManageChargeAmount(previousCoinAmount, currentCoinAmount);
-    this.replaceManageCoin(previousCoinEntries, currentCoinEnries);
-    this.render.reRenderVendingMachineManageMenuTemplate(this.vendingMachineManageMenuTemplate);
-  };
-
   onClickCoinReturnButton = () => {
     const $coinReturnButton = document.querySelector(DOM.$COIN_RETURN_BUTTON);
     $coinReturnButton.addEventListener(EVENT.CLICK, () => {
       const previousCoinsHash = { ...this.coins.getCoinsHash() };
       const previousCoinAmount = this.coins.getCoinAmount();
+
       if (previousCoinAmount === NUMBER.ZERO) {
         this.render.alertMessage(ERROR_MESSAGE.CHARGE_COIN_FIRST);
+
         return;
       }
 
-      this.hasReturnCoin();
-      this.renderPurchaseChargeAmount();
-      this.reRenderMachineManageMenu(Object.entries(previousCoinsHash), previousCoinAmount);
+      const setCoinReturn = new SetCoinReturn(this.render, this.coins, this.vendingMachine);
+      setCoinReturn.main(previousCoinsHash, previousCoinAmount);
     });
   };
